requirements.txt
absl-py==0.11.0
albumentations==0.5.1
autopep8==1.5.4
cachetools==4.1.1
certifi==2020.6.20
chardet==3.0.4
cycler==0.10.0
dataclasses==0.6
decorator==4.4.2
efficientnet-pytorch==0.6.3
fsspec==0.8.4
future==0.18.2
google-auth==1.23.0
google-auth-oauthlib==0.4.2
grpcio==1.33.2
idna==2.10
imageio==2.9.0
imgaug==0.4.0
install==1.3.4
kiwisolver==1.3.1
Markdown==3.3.3
matplotlib==3.3.2
munch==2.5.0
networkx==2.5
numpy==1.19.4
oauthlib==3.1.0
opencv-python==4.4.0.46
opencv-python-headless==4.4.0.46
pandas==1.1.4
Pillow==8.0.1
pretrainedmodels==0.7.4
protobuf==3.13.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.6.0
pyparsing==2.4.7
python-dateutil==2.8.1
python-dotenv==0.15.0
pytorch-lightning==1.0.0
pytz==2020.4
PyWavelets==1.1.1
PyYAML==5.3.1
requests==2.24.0
requests-oauthlib==1.3.0
rsa==4.6
scikit-image==0.17.2
scipy==1.5.4
segmentation-models-pytorch==0.1.2
Shapely==1.7.1
six==1.15.0
tensorboard==2.3.0
tensorboard-plugin-wit==1.7.0
tifffile==2020.10.1
timm==0.1.20
toml==0.10.2
torch==1.7.0
torchvision==0.8.1
tqdm==4.51.0
typing-extensions==3.7.4.3
urllib3==1.25.11
Werkzeug==1.0.1
